// rs-x-ui/lib/dom-element/dom-element.interface.ts
import { ITracker } from '@rs-x/core';
import { IPosition } from '../align-to/position.interface';


export interface IDomElement<T extends object = HTMLElement> {
   id: string;
   nativeElement: T;
   left: number;
   top: number;
   height: number;
   width: number;
   readonly offset: IPosition;
   readonly outerOffset: IPosition;
   marginLeft: number;
   marginRight: number;
   marginTop: number;
   marginBottom: number;
   paddingLeft: number;
   paddingRight: number;
   paddingTop: number;
   paddingBottom: number;
   borderLeft: number;
   borderRight: number;
   borderBottom: number;
   borderTop: number;
   innerHeight: number;
   innerWidth: number;
   outerWidthWithMargin: number;
   outerHeight: number;
   outerHeightWithMargin: number;
   outerWidth: number;
   scrollTop: number;
   scrollLeft: number;
   firstChild: IDomElement<T>;
   children: IDomElement<T>[];
   hasXOverflow: boolean;
   hasYOverflow: boolean;
   isVisible: boolean;
   innerHtml: string;
   enabled: boolean;
   mousedown: ITracker<MouseEvent, unknown>;
   mousemove: ITracker<MouseEvent, unknown>;
   mouseover: ITracker<MouseEvent, unknown>;
   mouseout: ITracker<MouseEvent, unknown>;
   addClass(className: string): void;
   removeClass(className: string): void;
   hasClass(className: string): boolean;
   find(selector: string): IDomElement | null;
   findAll(selector: string): IDomElement[];
   getParent(selector?: string): IDomElement;
   getCssAttributeValue(cssAttributeName: string): string;
   setCssAttributeValues(c: Record<string, string>): void;
   isOutSide(selector: unknown, excludeSelectors?: string[]): boolean;
   getAttribute(name: string): string | null;
   setAttribute(name: string, value: string): void;
   scrollLeftToEnd(): void;
   scrollTopToEnd(): void;
   addChild(child: IDomElement): void;
   remove(): void;
   clearContent(): void;
}